pyproject.toml
[project]
name = "nte-login"
version = "0.1.0"
description = "NTEUID 外置登录服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.130.0",
    "uvicorn[standard]>=0.40.0",
    "jinja2>=3.1.6",
    "httpx>=0.28.0",
    "loguru>=0.7.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.10.0",
    "websockets>=14.0",
]

[dependency-groups]
dev = ["ruff>=0.8.0"]

[project.scripts]
start = "ntelogin.main:main"

[tool.setuptools]
packages = ["ntelogin"]

[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I"]

[tool.pyright]
include = ["ntelogin"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
